requirements.txt
alembic==1.11.2
Babel==2.12.1
bcrypt==4.0.1
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.1
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.6
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
wincertstore==0.2
WTForms==3.0.1
zipp==3.16.2
